Reduce PR100 redaction duplication
