+++ story-generator/src/app/app.ts
@@ -925,7 +925,7 @@ ${chapters}
   }
 
   private isAsciiLetterOrDigit(char: string): boolean {
-    const code = char.charCodeAt(0);
+    const code = char.codePointAt(0) ?? 0;
     return (code >= 48 && code <= 57) || (code >= 65 && code <= 90) || (code >= 97 && code <= 122);
   }
